[CL-1546] Updates from code review
