+++ front/app/containers/Admin/formBuilder/components/FormBuilderSettings/utils.tsx
@@ -1,11 +1,17 @@
 import React from 'react';
+import { ICustomFieldInputType } from 'services/formCustomFields';
+import { Locale } from 'typings';
 
 // Components
 import ConfigMultiselectWithLocaleSwitcher from './ConfigMultiselectWithLocaleSwitcher';
 import LinearScaleSettings from './LinearScaleSettings';
 
 // Function to return additional settings based on input type
-export function getAdditionalSettings(type: any, locales: any, index: any) {
+export function getAdditionalSettings(
+  type: ICustomFieldInputType,
+  locales: Locale[],
+  index: number
+) {
   switch (type) {
     case 'multiselect':
       return (
